add a crud function class with restcontroller

+++ src/main/java/com/example/demo/Service/SogneService.java
@@ -2,28 +2,35 @@ package com.example.demo.Service;
 
 import com.example.demo.Model.Kommune;
 import com.example.demo.Model.Sogne;
+import com.example.demo.Model.SogneDTO;
+import com.example.demo.Repository.KommuneRepository;
 import com.example.demo.Repository.SogneRepository;
-import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
 
 import java.util.*;
 
+@Service
 public class SogneService {
-     /*
-      void create();
-    long findKommuneById(long id);
-    List<Kommune> findAllKommune();
-    void update(long id);
-    void delete (long id);
-     */
 
     @Autowired
     SogneRepository sogneRepository;
 
+    @Autowired
+    KommuneRepository kommuneRepository;
+
     public Sogne create(Sogne sogne){
         return sogneRepository.save(sogne);
     }
-    public Set<Sogne> findAllKommune(){
+
+    public Sogne create(SogneDTO sogneDTO){
+        Optional<Kommune> kommune = kommuneRepository.findById(sogneDTO.getKommunekode());
+        return sogneRepository.save(new Sogne(sogneDTO, kommune.get()));
+    }
+
+    public Set<Sogne> findAll(){
         Set<Sogne> sogneSet = new HashSet<>();
         for(Sogne sogne: sogneRepository.findAll()){
             sogneSet.add(sogne);
@@ -31,7 +38,7 @@ public class SogneService {
         return sogneSet;
     }
 
-    public Sogne findKommuneById(long id){
+    public Sogne findById(long id){
         Optional<Sogne> sogneOptional = sogneRepository.findById(id);
         if (!sogneOptional.isPresent()){
             throw new RuntimeException("Sogne: " + id + " sogne blev ikke fundet");
@@ -40,21 +47,33 @@ public class SogneService {
         }
     }
 
-    public Sogne update(Sogne sogne){
-        return sogneRepository.save(sogne);
-    }
-
-    public void delete (long id){
-        sogneRepository.deleteById(id);
-    }
-
-    public boolean checkIfSogneExist(long id) throws NotFoundException {
-        boolean checkIfExist = sogneRepository.existsById(id);
-        boolean status = false;
-        if (!checkIfExist){
-            throw new NotFoundException("Sogne dosent exist : " + id);
-        } else {
-            return status = true;
+    public Sogne update(SogneDTO sogneDTO) throws Exception {
+        Iterator<Sogne> iterator = sogneRepository.findAll().iterator();
+        while (iterator.hasNext()) {
+            Sogne sogne = iterator.next();
+            if(sogneDTO.getNavn() == sogne.getNavn()){
+                return sogneRepository.save(sogne);
+            }
         }
+        throw new Exception("Kan ikke updatere");
     }
+
+    public ResponseEntity<String> delete (long id){
+        //sogneRepository.deleteById(id);
+
+
+        Optional<Sogne> optionalParish = sogneRepository.findById(id);
+        if (optionalParish.isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("{ 'msg' : 'parish_id " + id + " not found'}");
+        }
+
+        Sogne sogne = optionalParish.get();
+
+
+        sogneRepository.save(sogne);
+        sogneRepository.deleteById(id);
+
+        return ResponseEntity.status(HttpStatus.OK).body("{ 'msg' : 'deleted'}");
+    }
+
 }
